// src/model/courseLogsMetricsRequestDtoFieldsItem.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * UBUMonitor-Analytics
 * API documentation for UBUMonitor Analytics
 * OpenAPI spec version: 1.0.0
 */

export type CourseLogsMetricsRequestDtoFieldsItem = typeof CourseLogsMetricsRequestDtoFieldsItem[keyof typeof CourseLogsMetricsRequestDtoFieldsItem];


export const CourseLogsMetricsRequestDtoFieldsItem = {
  COURSE_ID: 'COURSE_ID',
  USER_ID: 'USER_ID',
  MODULE_ID: 'MODULE_ID',
  COMPONENT_ID: 'COMPONENT_ID',
  EVENT_ID: 'EVENT_ID',
  ORIGIN_ID: 'ORIGIN_ID',
  IP_ADDRESS: 'IP_ADDRESS',
  USER_FULL_NAME: 'USER_FULL_NAME',
  MODULE_NAME: 'MODULE_NAME',
  COMPONENT_NAME: 'COMPONENT_NAME',
  EVENT_NAME: 'EVENT_NAME',
  ORIGIN_NAME: 'ORIGIN_NAME',
  SECTION_ID: 'SECTION_ID',
  SECTION_NAME: 'SECTION_NAME',
} as const;