// src/model/courseLogsMetricsRequestDtoGroupByItem.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * UBUMonitor-Analytics
 * API documentation for UBUMonitor Analytics
 * OpenAPI spec version: 1.0.0
 */

export type CourseLogsMetricsRequestDtoGroupByItem = typeof CourseLogsMetricsRequestDtoGroupByItem[keyof typeof CourseLogsMetricsRequestDtoGroupByItem];


export const CourseLogsMetricsRequestDtoGroupByItem = {
  USER_ID: 'USER_ID',
  MODULE_ID: 'MODULE_ID',
  COMPONENT_ID: 'COMPONENT_ID',
  EVENT_ID: 'EVENT_ID',
  ORIGIN_ID: 'ORIGIN_ID',
  IP_ADDRESS: 'IP_ADDRESS',
  SECTION_ID: 'SECTION_ID',
} as const;