// src/model/courseLogsMetricsResponseRowDto.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * UBUMonitor-Analytics
 * API documentation for UBUMonitor Analytics
 * OpenAPI spec version: 1.0.0
 */
import type { CourseLogsMetricsResponseRowComponentDto } from './courseLogsMetricsResponseRowComponentDto';
import type { CourseLogsMetricsResponseRowEventDto } from './courseLogsMetricsResponseRowEventDto';
import type { CourseLogsMetricsResponseRowModuleDto } from './courseLogsMetricsResponseRowModuleDto';
import type { CourseLogsMetricsResponseRowOriginDto } from './courseLogsMetricsResponseRowOriginDto';
import type { CourseLogsMetricsResponseRowUserDto } from './courseLogsMetricsResponseRowUserDto';

export interface CourseLogsMetricsResponseRowDto {
  timeBucket?: string;
  user?: CourseLogsMetricsResponseRowUserDto;
  module?: CourseLogsMetricsResponseRowModuleDto;
  component?: CourseLogsMetricsResponseRowComponentDto;
  event?: CourseLogsMetricsResponseRowEventDto;
  origin?: CourseLogsMetricsResponseRowOriginDto;
  ipAddress?: string;
  courseId?: number;
  sectionId?: number;
  sectionName?: string;
  value: number;
}